# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Crypto Investigation API",
    "description": "API contract specifications for cases, wallet tracing, risk, and alerts.",
    "version": "1.0.0"
  },
  "paths": {
    "/wallets/{id}/trace": {
      "get": {
        "summary": "Trace Wallet",
        "operationId": "trace_wallet_wallets__id__trace_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "depth",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5,
              "default": 2,
              "title": "Depth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletTraceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/wallets/{id}/risk": {
      "get": {
        "summary": "Get Wallet Risk",
        "operationId": "get_wallet_risk_wallets__id__risk_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletRiskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/wallets/{id}/vasp": {
      "get": {
        "summary": "Get Wallet Vasp",
        "operationId": "get_wallet_vasp_wallets__id__vasp_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WalletVASPResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cases": {
      "get": {
        "summary": "List Cases",
        "operationId": "list_cases_cases_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CaseResponse"
                  },
                  "type": "array",
                  "title": "Response List Cases Cases Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Case",
        "operationId": "create_case_cases_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reports/{case_id}": {
      "get": {
        "summary": "Generate Case Report",
        "operationId": "generate_case_report_reports__case_id__get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/alerts": {
      "get": {
        "summary": "Get Alerts",
        "operationId": "get_alerts_alerts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AlertResponse"
                  },
                  "type": "array",
                  "title": "Response Get Alerts Alerts Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AlertResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "wallet_address": {
            "type": "string",
            "title": "Wallet Address"
          },
          "rule_name": {
            "type": "string",
            "title": "Rule Name"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "wallet_address",
          "rule_name",
          "severity",
          "created_at"
        ],
        "title": "AlertResponse"
      },
      "CaseCreateRequest": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "CaseCreateRequest"
      },
      "CaseReportResponse": {
        "properties": {
          "case_id": {
            "type": "integer",
            "title": "Case Id"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "total_tainted_value": {
            "type": "number",
            "title": "Total Tainted Value"
          },
          "involved_wallets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Involved Wallets"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          }
        },
        "type": "object",
        "required": [
          "case_id",
          "summary",
          "total_tainted_value",
          "involved_wallets",
          "generated_at"
        ],
        "title": "CaseReportResponse"
      },
      "CaseResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/NormalizedTransaction"
            },
            "type": "array",
            "title": "Transactions",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "status",
          "created_at"
        ],
        "title": "CaseResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "NormalizedTransaction": {
        "properties": {
          "from": {
            "type": "string",
            "title": "From"
          },
          "to": {
            "type": "string",
            "title": "To"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "asset": {
            "type": "string",
            "title": "Asset"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "tx_hash": {
            "type": "string",
            "title": "Tx Hash"
          },
          "chain": {
            "type": "string",
            "title": "Chain"
          },
          "fee": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fee",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "from",
          "to",
          "amount",
          "asset",
          "timestamp",
          "tx_hash",
          "chain"
        ],
        "title": "NormalizedTransaction"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WalletRiskResponse": {
        "properties": {
          "wallet_address": {
            "type": "string",
            "title": "Wallet Address"
          },
          "risk_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Risk Score"
          },
          "risk_level": {
            "type": "string",
            "title": "Risk Level"
          },
          "flagged_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Flagged Reasons"
          }
        },
        "type": "object",
        "required": [
          "wallet_address",
          "risk_score",
          "risk_level",
          "flagged_reasons"
        ],
        "title": "WalletRiskResponse"
      },
      "WalletTraceResponse": {
        "properties": {
          "root_wallet": {
            "type": "string",
            "title": "Root Wallet"
          },
          "depth": {
            "type": "integer",
            "title": "Depth"
          },
          "total_volume": {
            "type": "number",
            "title": "Total Volume"
          },
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/NormalizedTransaction"
            },
            "type": "array",
            "title": "Transactions"
          }
        },
        "type": "object",
        "required": [
          "root_wallet",
          "depth",
          "total_volume",
          "transactions"
        ],
        "title": "WalletTraceResponse"
      },
      "WalletVASPResponse": {
        "properties": {
          "wallet_address": {
            "type": "string",
            "title": "Wallet Address"
          },
          "vasp_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vasp Name"
          },
          "is_sanctioned": {
            "type": "boolean",
            "title": "Is Sanctioned",
            "default": false
          },
          "category": {
            "type": "string",
            "title": "Category"
          }
        },
        "type": "object",
        "required": [
          "wallet_address",
          "category"
        ],
        "title": "WalletVASPResponse"
      }
    }
  }
}